Move inline tests into dedicated modules
